+++ tests/e2e/specs/rest-api/grouped-product.test.js
@@ -1,0 +1,6 @@
+/*
+ * Internal dependencies
+ */
+const { runGroupedProductAPITest } = require( '@woocommerce/e2e-core-tests' );
+
+runGroupedProductAPITest();